requirements.txt
openai
sqlalchemy
pymysql
python-dotenv
pandas
fastapi 
uvicorn
cryptography